+++ client/styles/globals.css
@@ -1,4 +1,4 @@
-:root {
+ :root {
   --max-width: 1100px;
   --border-radius: 12px;
   --font-mono: ui-monospace, Menlo, Monaco, 'Cascadia Mono', 'Segoe UI Mono',
@@ -100,8 +100,8 @@ a {
   text-decoration: none;
 }
 
-@media (prefers-color-scheme: dark) {
+ @media (prefers-color-scheme: dark) {
   html {
     color-scheme: dark;
   }
-}
+}
